QQuickImage: Fix an odd 1px offset on 'fillMode: PreserveAspectFit'
We used to ceil x and y offsets on default (AlignHCenter) PreserveAspectFit
image. Consider the follow code:

    width = 256;
    pixWidth = 255.99;
    xOffset = ceil((width - pixWidth) / 2);

It means that we use 1px offset to center an image that is 0.01px tighter
than its view. This actually leads to painting outside of the bounding rect.

Use simple integer division to calculate x and y offsets.

Task-number: QTBUG-79011
Change-Id: I76f2fe3361b820c48202f8f872e2ddb58a65016e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>