Fix updating of text nodes in QQuickTextEdit
The update algorithm wasn't correctly removing all node that
requires removing, as the startPosition is not always up to
date when deleting lines of code.

Instead, figure out the first Node after the changed region that
is clean and keep that one as a reference.

Amends 560a1991ac4524ff16352da23a2b54d717548f33.

Fixes: QTBUG-74745
Change-Id: I4a2c5bd7a673af5cad1850e3a5b703f9554cd7e6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Igor Bugaev <freedbrt@gmail.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>