ParentChange: Restore old position correctly
ParentChange does floating point math in doChange to calculate the
correct position. Unfortunately, while doing the conversion, we
accumulate rounding errors in the presence of rotations.

Those can lead to visual glitches, as observed in QTBUG-68176.

This patch avoids the issue by storing the old values and resetting to
them in restore.

Fixes: QTBUG-68176
Change-Id: I6ebe1ccbc601838aa664cdc723e0cd58c95e785a
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>