Make ScopedValue's CHECK_EXCEPTION also check isInterrupted
While there should be no way to generate an infinite loop from any of
the affected methods, you can certainly generate a really long loop, for
example with Array(1E9).join(). We should be able to interrupt this.

Also, the various call()s could return with isInterrupted set. We should
respect that and immediately return.

Fixes: QTBUG-78955
Change-Id: I7e18b24db0bf39df03134027b2b5dba452ac7c1c
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>