Fix a layout bug caused by a delegate item that was moved by the user
The position of the first item in the list of visualItems was used to
know how to layout the rest of the visual items.
However, this did not work if the first item was actually moved (e.g.
due to a DnD operation). We therefore store the position of the first
visual item after each time we arrange it, and use that as a basis on
where to start layouting from.

Task-number: QTBUG-78076
Change-Id: I837f5b7d61a13d98d23287685c6fd66817360906
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>