Android: Fix crash in release mode
When using the "auto" keyword, we would keep a reference to the QStringBuilder,
instead of converting it to a QString. This object would in turn keep
references to the pluginName string, which the compiler would delete,
and we would get a crash later on, when the actual conversion took place.

So this is a sneaky compiler bug and the work-around is to explicitly
convert the QStringBuilder to a QString right away.

Fixes: QTBUG-79230
Change-Id: I0759645c84d6f995f26063ea098cdaea61e924ab
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>