Fix broken scaling with high dpi + rhi + thread loop
The arguments to renderSceneGraph() were incorrect in the threaded
render loop, unlike basic. Fix it up to follow what we do in the
basic loop. (The first argument is the QWindow size (logical). The
second is only used when rendering via the rhi, and is the output
surface/layer size, in pixels. This will get simplified in Qt 6.)

This fixes broken rendering with Metal and the threaded render loop
for windows with dpr > 1.

Change-Id: I427ba5def8cc34900e6fe650e1006ca5f0fa90a4
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>