Fix assertion with certain Text items when rendering with Metal
Caused by a -1 vs. 0 discrepancy between the QRhi and the OpenGL
glyph caches. width() and height() are expected to return 0 for
an empty image, but this we forgot to honor in the QRhi version.
That returned the width or height of a QSize, and those are
-1 when default constructed. (there are good reasons for that,
it just happens to not match the non-QSize-based code path here)

Change-Id: Idda0d2f77430abea79d4004b765077032e06954b
Fixes: QTBUG-79533
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>