Adapt test case to avoid null characters in directory names
The file system engine in QtCore is stricter now about these things. In
the particular example we have here, the first character of the imported
directory was a null character, which would previously lead to the whole
import being ignored. We cannot do such nasty things anymore. Therefore,
comment the import out, in order to retain the old behavior.

Fixes: QTBUG-79562
Change-Id: I04917bf7db32e9d09e5919555740a263af2d1795
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>