Non-modal popups shouldn't set isTabFence
When a popup is not modal, all the application's main content can still
be interacted with, therefore should still be possible to navigate it
with tab. This issue is particularly evident with Drawer, also commonly
used as a sidebar, in which tab navigation should always work.

Fixes: QTBUG-79501
Change-Id: I0e8ad2e1abe57b9617c6316efbfe2f296c91b592
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3358362e61ef96b3922a438d0806561285c49230)