Take a copy of the children list before emitting the destroyingItem signal
clear() emits a signal for each item that is being destroyed. Unfortunately
that signal can be connected to some other place that already removes
the item indirectly. In that case, we would not correctly emit the
destroyed signal for all items, leading to errors later on.

Fixes a test failure in Qt Quick Controls 2

Change-Id: I3aff75d9263badd9f87883610c7a00a94ee823d3
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>