Fix warning about non-relative paths in qmldir files
Permit the use of resource urls in qmldir files to allow for the
use-case of a plugin including its .qml files in resources and
optionally compiling them ahead of time. When the resources are bundled
in the plugin, qmlplugindump might or might not work.

Task-number: QTBUG-48809
Change-Id: Icb331c7575f26316b5c2bcc604b9c6d793977a9f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit f15c2c22cb677677ccf8aacb2603fcf774d13584)