manual test, tableview: Ensure we modify the right model
An old, unused, model was just lying around in the code. And the
buttons in the test was modifying it. But after introducing the
syncView concept in 5.14, all the TableViews in the test got their own
separate model. So fix up the test to work as it should with the new setup.

Change-Id: Ic3e6e3013186ef1c2998566b84b3b94e809296e8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>