Share pipeline state objects between layer renderers
...as long as the renderpass descriptors are compatible.

Relies on the new QRhiRenderPassDescriptor::isCompatible() function.

In the nodetypes_ng manual test this brings down the number of pipelines
created from 16 to 8 in the layer test scene (because most layers use
the same materials there, so many pipelines can be reused between them -
previously this was prevented by the comparison-by-pointer for the
QRhiRenderPassDescriptors). Additionally, any application with more than one
layer (Item.layer enabled or ShaderEffect) in the scene will benefit from this.

Task-number: QTBUG-80318
Change-Id: Id7437a3baac31891a231643d4787da7186ac107e
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>