Make the initialization failure dialog sensible with QRhi
So on Windows one now gets a message box with a reasonable message,
instead of the OpenGL nonsense. Then the application closes when pressing
Abort etc.

On other platforms there is a qFatal, printing the same message.

Involves simplifying the OpenGL version a bit since passing isES type of
flags through multiple layers is not justified here.

Task-number: QTBUG-80365
Change-Id: Ie3ea1e9395a283f7e95eda78c1d3894797ff0acf
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>