Make inputmask 'X' mask character require non-blank input
The implementation for the X mask would accept any printable
character. Since that includes the blank character, there was no
difference in behavior between the requiring X and optional x: both
would allow the input to be unset, i.e. blank.

This change should be seen in conjunction with the doc improvement
qtbase:da0af1e and the corresponding change to the QLineEdit widget,
where the same input mask handling code is duplicated.

This patch series also concludes the old task questioning whether
text() should strip blanks even when they are valid input (answer:
yes, continue to do that), and so a couple of XFAILS was removed from
the autotest.

[ChangeLog][QtQuick][TextInput] Inputmask X character now requires non-blank input.

Task-number: QTBUG-76320
Change-Id: I606ae04259f29f748b3ce604048c6eb6f10c2ff9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>