Clear RequestRepaint bit set in sync() if we are going to render
...because if we are going to go on doing a frame then the repaint request
is fulfilled. Leaving the bit set would cause the render thread's main loop
to start another frame right after the current one is queued, which is kind
of bad in certain scheduling and blocking cases because the main thread
may end up starving due to the unnecessary extra (and sync-less) frame.

Task-number: QTBUG-80498
Change-Id: I6c9a8f60e5f49d771d7d284bb7399eadbde33b8c
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>