Don't allow dragging a ListView through a floating header or footer
If the header or footer positioning is Overlay or PullBack, the list
delegates appear to scroll underneath it.  The header or footer can
contain interactive content.  If a mouse or touch press happens to
"fall through" that, it should not be possible to drag the ListView
contents from there.

[ChangeLog][QtQuick][ListView] ListView no longer allows the user to
press on an Overlay or PullBack header or footer and start scrolling,
but only on the content delegates.

Fixes: QTBUG-74046
Change-Id: I4d06c789286be2691d098aeebb10a57b2a335744
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>