QQuickTableView: set empty content size when table is empty
From before we would bail out early from the rebuild process if we
detected an empty table. A result from this is that we left
both contentWidth and contentHeight unchanged.

This patch will set an empty content size when the table is
empty. The effect will be that the user cannot flick the view
around based on the old size.

Fixes: QTBUG-80505
Change-Id: I3ac080476269fd5906ce79fa007eabb59b5ff4b1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>