QQuickTableView: ensure we release items in the old model and not the new
As it stood, we would wait to release loaded items until we started the
rebuild process, if the old model was a DelegateModel. But at that time,
the model would alread have been changed, so we would release the items
by calling out to the wrong model.

This patch will ensure that we always release the items immediately when
syncing the model, which will also cover the case when the model is a
DelegateModel.

Fixes: QTBUG-80570
Change-Id: I1b06011f4795727d04d9cd8c20381f65552b8fe8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>